introapi/pyproject.toml
[tool.poetry]
name = "introapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django-ninja = "^0.21.0"
Django = "^4.1.7"
django-cors-headers = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
